% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{read_reports}
\alias{read_reports}
\title{Reads in the reports data from pins board}
\usage{
read_reports(choice = NULL)
}
\arguments{
\item{choice}{if `NULL`, reads in all of the reports and returns a list.
Otherwise, one of "overall" or a province code ("AB", "BC", etc.).}
}
\value{
A single data frame, or a named list of data frames (if `choice`
  is `NULL`).
}
\description{
Reads in the reports data from pins board
}
