% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{read_summary}
\alias{read_summary}
\title{Reads in the summary data from pins board}
\usage{
read_summary(split = c("overall", "province"))
}
\arguments{
\item{split}{One of "overall" (aggregated counts across Canada) or "province"
(splits counts by province/territory).}
}
\value{
A data frame.
}
\description{
Reads in the summary data from pins board
}
