% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valueBoxCustom}
\alias{valueBoxCustom}
\title{A custom value box}
\usage{
valueBoxCustom(
  value,
  subtitle,
  icon = NULL,
  color = "white",
  background = "aqua",
  width = 4,
  href = NULL
)
}
\arguments{
\item{value}{The value to display in the box. Usually a number or short text.}

\item{subtitle}{Subtitle text.}

\item{icon}{An icon tag, created by \code{\link[shiny]{icon}}.}

\item{color}{A color for the box.}

\item{href}{An optional URL to link to.}
}
\description{
A slightly adapted `shinydashboard::valueBox` function which allows for
other colors (besides the `shinydashboard::validColors`).
Credit goes to [this user on the RStudio forum](https://community.rstudio.com/t/shinydashboard-custom-box-colors-to-match-brand/14147/5).
}
