% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{write_data}
\alias{write_data}
\title{Reads in the reports data from pins board}
\usage{
write_data(data, name)
}
\arguments{
\item{data}{The data to save to the pin board.}

\item{name}{The name to give the data.}
}
\description{
Reads in the reports data from pins board
}
